Extensive refactoring of tests for query modules.
